+++ polyfill/changelog.md
@@ -1,3 +1,15 @@
+## 1.2.0 (2019-04-18)
+
+* Add UX test cases for verifying the distance function [(#154)](https://github.com/WICG/spatial-navigation/pull/154)
+* Improve the distance function depending on [the spec update](https://github.com/w3c/csswg-drafts/pull/3755) [(#165)](https://github.com/WICG/spatial-navigation/pull/165)
+* Fix the bug for preventing navigation events
+* Remove `navbeforescroll` event
+* Add `spatial-navigation-action` CSS Property and its demo [(#171)](https://github.com/WICG/spatial-navigation/pull/171)
+* Add internal test cases for overall polyfill behavior
+* Add the feature about navigating from the spatial navigation starting point [(#169)](https://github.com/WICG/spatial-navigation/pull/169)
+* Optimize `navigate()` [(#172)](https://github.com/WICG/spatial-navigation/pull/172)
+* Fix the bug of handling non-focusable elements [(#174)](https://github.com/WICG/spatial-navigation/pull/174)
+
 ## 1.1.0 (2019-01-04)
 
 * Change the polyfill file name from spatnav-heuristic.js to spatial-navigation-polyfill.js